cleanup assets folder, delete pbrocks folder
