+++ assets/gather-blocks.php
@@ -1,0 +1,6 @@
+<?php
+
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
