+++ assets/package.json
@@ -1,12 +1,13 @@
 {
 	"name": "gatherpress",
 	"version": "1.0.0",
-	"description": "",
+	"description": "ES6 Assets for GatherPress including Blocks, Filters, and Plugins",
 	"main": "index.js",
 	"scripts": {
 		"build": "wp-scripts build",
 		"check-engines": "wp-scripts check-engines",
 		"check-licenses": "wp-scripts check-licenses",
+		"dev": "wp-scripts start",
 		"lint:css": "wp-scripts lint-style",
 		"lint:js": "wp-scripts lint-js",
 		"lint:pkg-json": "wp-scripts lint-pkg-json",
@@ -14,22 +15,30 @@
 		"test:e2e": "wp-scripts test-e2e",
 		"test:unit": "wp-scripts test-unit-js"
 	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/mauteri/gatherpress.git"
+	},
+	"keywords": [
+		"dates",
+		"events"
+	],
 	"author": "",
-	"license": "ISC",
+	"license": "GPL-2.0",
+	"bugs": {
+		"url": "https://github.com/mauteri/gatherpress/issues"
+	},
+	"homepage": "https://github.com/mauteri/gatherpress#readme",
 	"devDependencies": {
-		"@wordpress/scripts": "^7.1.3",
-		"bootstrap": "^4.4.1",
-		"copy-webpack-plugin": "^5.1.1",
-		"css-loader": "^3.4.2",
-		"file-loader": "^6.0.0",
+		"@wordpress/components": "^8.3.2",
+		"@wordpress/scripts": "^5.1.0",
+		"css-loader": "^3.2.0",
 		"ignore-emit-webpack-plugin": "^2.0.2",
-		"jquery": "^3.4.1",
-		"mini-css-extract-plugin": "^0.9.0",
+		"mini-css-extract-plugin": "^0.8.0",
 		"node-sass": "^4.13.1",
 		"path": "^0.12.7",
-		"popper.js": "^1.16.1",
 		"postcss-loader": "^3.0.0",
 		"postcss-preset-env": "^6.7.0",
-		"sass-loader": "^8.0.2"
+		"sass-loader": "^7.1.0"
 	}
 }
