--- assets/src/js/blocks.js
@@ -1,11 +1,0 @@
-
-import './blocks/event-datetime';
-// import './filters';
-
-// svg code
-import gatherPressIcon from './../../images/gatherpress-icon';
-
-// alter the icon slot
-wp.blocks.updateCategory( 'gatherpress', {
-    icon: gatherPressIcon
-} );
