--- assets/src/js/blocks/index.js
@@ -1,3 +1,0 @@
-import './event-datetime';
-// import './time-sidebar';
-// import './sample';
