--- assets/src/scss/style.scss
@@ -1,1 +1,0 @@
-@import 'bootstrap';
