+++ functions.php
@@ -19,7 +19,7 @@ define( 'GATHERPRESS_THEME_VERSION', wp_get_theme()->get( 'Version' ) );
 require_once GATHERPRESS_CORE_PATH . '/inc/helpers/autoloader.php';
 require_once GATHERPRESS_CORE_PATH . '/plugins/jetpack-tweaks/css-sanitization.php';
 require_once GATHERPRESS_CORE_PATH . '/plugins/wordcamp-remote-css/bootstrap.php';
-require_once GATHERPRESS_CORE_PATH . '/pbrocks/gather-blocks.php';
+require_once GATHERPRESS_CORE_PATH . '/assets/gather-blocks.php';
 
 // Kick things off!
 \GatherPress\Inc\Setup::get_instance();
