--- pbrocks/src/index.js
@@ -1,6 +1,0 @@
-/**
- * Internal dependencies
- */
-
-import './blocks';
-import './filters';
